[FEATURE] Added allowLanguageSynchronization to TCA
Added allowLanguageSynchronization to TCA of nearly all available
fields. An editor can now select, whether to use the default value
of the original translation or to use an own value.

This feature is only available in TYPO3 8.7+. I decided not to
implement "l10n_mode=mergeIfNotBlank" as a fallback for TYPO3 7.6
due to known problems with this setting
(see https://forge.typo3.org/issues/79658)

Closes #519