[FEATURE] Add JSON-LD output for events
Google need structured content to "understand" the meaning of content. This structure can be provided via
the Google Search Console or by utilizing a JSON-LD schema. In this case the event schema which is described in detail
here https://developers.google.com/search/docs/data-types/event

This PR uses the first image (if images available).

Closes #572