[BUGFIX] Added option to control enableFields setting in eventDemand
The `eventDemand` object has been extended, so the `enableFields` setting
for queries using `findDemanded` can be controlled. Additionally the
event model has been extended with getter/setters for the properties
hidden, starttime and endtime.

The list action of the administrationController now by default sets
the `enableFields` setting to false, so the event module also will show
events that are hidden ar have a start- or endtime.

Refs #805