[!!!][TASK] Removed language and workspace support for registrations
In order to fix problems with language labels and translations for
registrations record, language- and workspace-support has been removed
for the following tables:

* tx_sfeventmgt_domain_model_registration
* tx_sfeventmgt_domain_model_registration_fieldvalue

Extbase will still do a language overlay for all records that are
related to a registration (e.g. locations or events), so it is possible
to show registrations using a plugin on pages with multiple languages
(e.g. like the User Registrations plugin)

As a positive side effect, the performance of registration inline records
in the event record increased, since no language lookup need to be
processed.

This change can be considered as breaking, when the extension has been
extended and it is required that registrations have language related
fields. Since the language fields are not required and the extension
was not fully working on multilingual sites (see #821), this change
is added in a minor release.

Closes: #821
Refs: #819