[TASK] Add slug field for sys_category
This change adds the field "slug" to sys_category so it
is possible to create routes for the list view with
overwriteDemand/category as parameter.

Since ext:sf_event_mgt uses the same field name for the
slug as ext:news does, TYPO3 will only add the field once
if both extensions are installed.

Refs #835