[BUGFIX] Exclude `disableOverrideDemand` in dataHandler hook
A dataHandler hook ansures, that empty flexform values do not
get saved to the database. This is however problematic for
flexform fields with a default value of `1`, since the hook
removed the field from flexform when unchecked. After saving,
the flexform will then again show the default value, although
it is not set.

Refs #1148