[TASK] Check for existing event in checkConfirmRegistration()
When a user registers to an event which has access restrictions,
the event confirmation will fail, if the user is not logged in.
Currently, an exception is thrown, since the extension tries to
send an email while confirming.

This change adds an additional check to checkConfirmRegistration(),
so the confirmation is considered failed, if the registration has
no or no accessible event.

Refs #1155