[!!!][FEATURE] Introduce EventEvaluationService
To reduce logic in EventController, the following functions
have been moved to an own service:

* evaluateSingleEventSetting
* evaluateIsShortcutSetting
* evaluateEventPreviewSetting
* checkPidOfEventRecord

This additionally allows easier testing of the extracted
logic.

As a result or the refactoring, the PSR-14 event call was
moved to `EventEvaluationService`. This results in a
breaking change, because the `EventController` is now
not available via the PSR-14 event any more.

Closes #1238